+++ main_test.go
@@ -2,9 +2,18 @@ package main
 
 import (
 	"flag"
+	"fmt"
 	"io/ioutil"
+	"math"
 	"os"
+	"os/signal"
+	"runtime"
+	"strings"
+	"syscall"
 	"testing"
+	"time"
+
+	"github.com/mackerelio/mackerel-agent/command"
 )
 
 func TestParseFlags(t *testing.T) {
@@ -81,5 +90,67 @@ func TestParseFlagsRunOnce(t *testing.T) {
 
 	if otherOptions.runOnce == false {
 		t.Error("with -once args, RunOnce should be true")
+	}
+}
+
+func TestCreateAndRemovePidFile(t *testing.T) {
+	file, err := ioutil.TempFile("", "")
+	if err != nil {
+		t.Errorf("failed to create tmpfile, %s", err)
+	}
+	fpath := file.Name()
+	defer os.Remove(fpath)
+
+	err = createPidFile(fpath)
+	if err != nil {
+		t.Errorf("pid file should be created but, %s", err)
+	}
+
+	if runtime.GOOS != "windows" {
+		if err := createPidFile(fpath); err == nil || !strings.HasPrefix(err.Error(), "Pidfile found, try stopping another running mackerel-agent or delete") {
+			t.Errorf("creating pid file should be failed when the running process exists, %s", err)
+		}
+	}
+
+	removePidFile(fpath)
+	if err := createPidFile(fpath); err != nil {
+		t.Errorf("pid file should be created but, %s", err)
+	}
+
+	removePidFile(fpath)
+	ioutil.WriteFile(fpath, []byte(fmt.Sprint(math.MaxInt32)), 0644)
+	if err := createPidFile(fpath); err != nil {
+		t.Errorf("old pid file should be ignored and new pid file should be created but, %s", err)
+	}
+}
+
+func TestSignalHandler(t *testing.T) {
+	ctx := &command.Context{}
+	termCh := make(chan struct{})
+	c := make(chan os.Signal, 1)
+	signal.Notify(c, os.Interrupt, syscall.SIGTERM, syscall.SIGQUIT, syscall.SIGHUP)
+	go signalHandler(c, ctx, termCh)
+
+	resultCh := make(chan int)
+
+	maxTerminatingInterval = 100 * time.Millisecond
+	c <- os.Interrupt
+	c <- os.Interrupt
+
+	go func() {
+		<-termCh
+		<-termCh
+		<-termCh
+		<-termCh
+		resultCh <- 0
+	}()
+
+	go func() {
+		time.Sleep(time.Second)
+		resultCh <- 1
+	}()
+
+	if r := <-resultCh; r != 0 {
+		t.Errorf("Something went wrong")
 	}
 }