+++ tool/cover.sh
@@ -1,4 +1,4 @@
-#/bin/sh
+#!/bin/sh
 # Calculate coverage
 # run this and you can see coverage reports as following
 #   % go tool cover -html .profile.cov
@@ -8,10 +8,11 @@ prof=${1:-".profile.cov"}
 echo "mode: count" > $prof
 
 cleanup() {
+  retval=$?
   if [ $tmpprof != "" ] && [ -f $tmpprof ]; then
     rm -f $tmpprof
   fi
-  exit
+  exit $retval
 }
 trap cleanup INT QUIT TERM EXIT